address review: med-buzz off-switch, threshold refresh, wind-down quiet math
- clear armed MedBuzzer slots when medsEnabled is switched OFF (the OS
  slots already cancelled; the strap buzz kept buzzing — inconsistent).
  Preserve only on a failed schedule read while enabled, matching the
  scheduler's own rule.
- DeviceAlerts.refreshThreshold(): _restore is memoised, so a threshold
  change in Settings would not apply until the next process create. Now
  pushed through AppState on save, riding the alert queue.
- windDownSlot: wrap the pre-bedtime offset across midnight (a 00:20
  bedtime wants its wind-down at 23:35 the same evening, not a disabled
  feature); cap to half an hour before quiet opens and refuse if still
  inside the window — OS schedules never pass through shouldFireOs, so
  quiet hours must be applied here (check-in's own precedent). Drop
  skipToday so opting in mid-evening still gets tonight's slot.
- clamp() returns num; toInt() before it reaches int fields/prefs.

+++ lib/notify/notification_prefs.dart
@@ -181,9 +181,9 @@ class NotificationPrefs {
       movementEnabled: p.getBool(_kMovement) ?? false,
       medsEnabled: p.getBool(_kMeds) ?? false,
       checkInEnabled: p.getBool(_kCheckIn) ?? false,
-      batteryAlertPct:
-          (p.getInt(_kBatteryPct) ?? batteryPctDefault)
-              .clamp(batteryPctMin, batteryPctMax),
+      batteryAlertPct: ((p.getInt(_kBatteryPct) ?? batteryPctDefault)
+              .clamp(batteryPctMin, batteryPctMax))
+          .toInt(),
       stepGoalEnabled: p.getBool(_kStepGoal) ?? true,
       windDownEnabled: p.getBool(_kWindDown) ?? false,
     );
@@ -206,7 +206,7 @@ class NotificationPrefs {
     await p.setBool(_kMeds, medsEnabled);
     await p.setBool(_kCheckIn, checkInEnabled);
     await p.setInt(
-        _kBatteryPct, batteryAlertPct.clamp(batteryPctMin, batteryPctMax));
+        _kBatteryPct, batteryAlertPct.clamp(batteryPctMin, batteryPctMax).toInt());
     await p.setBool(_kStepGoal, stepGoalEnabled);
     await p.setBool(_kWindDown, windDownEnabled);
   }
